pyproject.toml
[build-system]
requires = ["scikit-build-core", "pybind11"]
build-backend = "scikit_build_core.build"


[project]
name = "emodlib"
version = "0.0.2"
description="A library of demographic, within-host, and contact-network algorithms for disease modeling"
readme = "README.md"
authors = [
  { name = "Edward Wenger", email = "wenger.edward@gmail.com" },
]
requires-python = ">=3.7"
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = ["pyyaml"]

[project.optional-dependencies]
test = ["pytest", "matplotlib", "numpy", "pandas", "xarray", "optuna", "ipywidgets", "plotly"]
dev = ["nox", "pytest"]


[tool.scikit-build]
wheel.expand-macos-universal-tags = true


[tool.cibuildwheel]
test-command = "pytest {project}/tests"
test-extras = ["test"]
test-skip = ["*universal2:arm64"]
build-verbosity = 1
